+++ apps/riscv-tests/isa/rv64uv/vmvnrr.c
@@ -4,7 +4,7 @@
 //
 // Author: Matteo Perotti
 //
-// For simplicity, this test depends on vl1r
+// For simplicity, this test depends on vl1r and vs1r
 
 #include "vector_macros.h"
 
@@ -12,21 +12,18 @@ uint64_t counter;
 
 // Maximum size: (VLEN/8 Bytes * (MAX_LMUL == 8)) = VLEN
 // Define VLEN before compiling me
-// #define VLEN 4096
+// #define VLEN VLEN
 uint8_t gold_vec_8b[VLEN];
-uint8_t zero_vec_8b[VLEN];
 uint8_t buf_vec_8b[VLEN];
 
-//////////
-// vs1r //
-//////////
+///////////
+// vmv1r //
+///////////
 
 // 1 whole register load
 void TEST_CASE1(void) {
   // Initialize a golden vector
   INIT_MEM_CNT(gold_vec_8b, VLEN / 8);
-  // Initialize a zero golden vector
-  INIT_MEM_ZEROES(zero_vec_8b, VLEN / 8);
   // Reserve space for a buffer in memory
   INIT_MEM_ZEROES(buf_vec_8b, VLEN / 8);
   // Set vl and vtype to super short values
@@ -35,24 +32,21 @@ void TEST_CASE1(void) {
   VCLEAR(v16);
   // Load a buffer from memory - whole register load
   asm volatile("vl1re8.v v16, (%0)" ::"r"(gold_vec_8b));
+  // Move the content to another register
+  asm volatile("vmv1r.v v1, v16");
   // Check that the whole register was loaded
-  asm volatile("vs1r.v v16, (%0)" ::"r"(buf_vec_8b));
+  asm volatile("vs1r.v v1, (%0)" ::"r"(buf_vec_8b));
   VMCMP(uint8_t, % hhu, 0, buf_vec_8b, gold_vec_8b, VLEN / 8);
-  // Check that the neighbour registers are okay
-  asm volatile("vs1r.v v17, (%0)" ::"r"(buf_vec_8b));
-  VMCMP(uint8_t, % hhu, 0, buf_vec_8b, zero_vec_8b, VLEN / 8);
 }
 
-//////////
-// vs2r //
-//////////
+///////////
+// vmv2r //
+///////////
 
 // 2 whole registers load
 void TEST_CASE2(void) {
   // Initialize a golden vector
   INIT_MEM_CNT(gold_vec_8b, VLEN / 4);
-  // Initialize a zero golden vector
-  INIT_MEM_ZEROES(zero_vec_8b, VLEN / 4);
   // Reserve space for a buffer in memory
   INIT_MEM_ZEROES(buf_vec_8b, VLEN / 4);
   // Set vl and vtype to super short values
@@ -61,24 +55,21 @@ void TEST_CASE2(void) {
   VCLEAR(v16);
   // Load a buffer from memory - whole register load
   asm volatile("vl2re8.v v16, (%0)" ::"r"(gold_vec_8b));
+  // Move the content to another register
+  asm volatile("vmv2r.v v2, v16");
   // Check that the whole register was loaded
-  asm volatile("vs2r.v v16, (%0)" ::"r"(buf_vec_8b));
-  VMCMP(uint8_t, % hhu, 1, buf_vec_8b, gold_vec_8b, VLEN / 8);
-  // Check that the neighbour registers are okay
-  asm volatile("vs2r.v v18, (%0)" ::"r"(buf_vec_8b));
-  VMCMP(uint8_t, % hhu, 1, buf_vec_8b, zero_vec_8b, VLEN / 8);
+  asm volatile("vs2r.v v2, (%0)" ::"r"(buf_vec_8b));
+  VMCMP(uint8_t, % hhu, 1, buf_vec_8b, gold_vec_8b, VLEN / 4);
 }
 
-//////////
-// vs4r //
-//////////
+///////////
+// vmv4r //
+///////////
 
 // 4 whole registers load
 void TEST_CASE3(void) {
   // Initialize a golden vector
   INIT_MEM_CNT(gold_vec_8b, VLEN / 2);
-  // Initialize a zero golden vector
-  INIT_MEM_ZEROES(zero_vec_8b, VLEN / 2);
   // Reserve space for a buffer in memory
   INIT_MEM_ZEROES(buf_vec_8b, VLEN / 2);
   // Set vl and vtype to super short values
@@ -87,24 +78,21 @@ void TEST_CASE3(void) {
   VCLEAR(v16);
   // Load a buffer from memory - whole register load
   asm volatile("vl4re8.v v16, (%0)" ::"r"(gold_vec_8b));
+  // Move the content to another register
+  asm volatile("vmv4r.v v4, v16");
   // Check that the whole register was loaded
-  asm volatile("vs4r.v v16, (%0)" ::"r"(buf_vec_8b));
-  VMCMP(uint8_t, % hhu, 2, buf_vec_8b, gold_vec_8b, VLEN / 8);
-  // Check that the neighbour registers are okay
-  asm volatile("vs4r.v v20, (%0)" ::"r"(buf_vec_8b));
-  VMCMP(uint8_t, % hhu, 2, buf_vec_8b, zero_vec_8b, VLEN / 8);
+  asm volatile("vs4r.v v4, (%0)" ::"r"(buf_vec_8b));
+  VMCMP(uint8_t, % hhu, 2, buf_vec_8b, gold_vec_8b, VLEN / 2);
 }
 
-//////////
-// vs8r //
-//////////
+///////////
+// vmv8r //
+///////////
 
 // 8 whole registers load
 void TEST_CASE4(void) {
   // Initialize a golden vector
   INIT_MEM_CNT(gold_vec_8b, VLEN);
-  // Initialize a zero golden vector
-  INIT_MEM_ZEROES(zero_vec_8b, VLEN);
   // Reserve space for a buffer in memory
   INIT_MEM_ZEROES(buf_vec_8b, VLEN);
   // Set vl and vtype to super short values
@@ -114,12 +102,11 @@ void TEST_CASE4(void) {
   VCLEAR(v24);
   // Load a buffer from memory - whole register load
   asm volatile("vl8re8.v v16, (%0)" ::"r"(gold_vec_8b));
+  // Move the content to another register
+  asm volatile("vmv8r.v v8, v16");
   // Check that the whole register was loaded
-  asm volatile("vs8r.v v16, (%0)" ::"r"(buf_vec_8b));
-  VMCMP(uint8_t, % hhu, 3, buf_vec_8b, gold_vec_8b, VLEN / 8);
-  // Check that the neighbour registers are okay
-  asm volatile("vs8r.v v24, (%0)" ::"r"(buf_vec_8b));
-  VMCMP(uint8_t, % hhu, 3, buf_vec_8b, zero_vec_8b, VLEN / 8);
+  asm volatile("vs8r.v v8, (%0)" ::"r"(buf_vec_8b));
+  VMCMP(uint8_t, % hhu, 3, buf_vec_8b, gold_vec_8b, VLEN);
 }
 
 ////////////
@@ -130,8 +117,6 @@ void TEST_CASE4(void) {
 void TEST_CASE5(void) {
   // Initialize a golden vector
   INIT_MEM_CNT(gold_vec_8b, VLEN / 8);
-  // Initialize a zero golden vector
-  INIT_MEM_ZEROES(zero_vec_8b, VLEN / 8);
   // Reserve space for a buffer in memory
   INIT_MEM_ZEROES(buf_vec_8b, VLEN / 8);
   // Set vl and vtype to super short values
@@ -140,12 +125,11 @@ void TEST_CASE5(void) {
   VCLEAR(v16);
   // Load a buffer from memory - whole register load
   asm volatile("vl1re8.v v16, (%0)" ::"r"(gold_vec_8b));
+  // Move the content to another register
+  asm volatile("vmv1r.v v1, v16");
   // Check that the whole register was loaded
-  asm volatile("vs1r.v v16, (%0)" ::"r"(buf_vec_8b));
+  asm volatile("vs1r.v v1, (%0)" ::"r"(buf_vec_8b));
   VMCMP(uint8_t, % hhu, 4, buf_vec_8b, gold_vec_8b, VLEN / 8);
-  // Check that the neighbour registers are okay
-  asm volatile("vs1r.v v17, (%0)" ::"r"(buf_vec_8b));
-  VMCMP(uint8_t, % hhu, 4, buf_vec_8b, zero_vec_8b, VLEN / 8);
 }
 
 int main(void) {