+++ CHANGELOG.md
@@ -10,6 +10,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 - Fix typo on the build instructions of the README
 - Fix Gnuplot installation on GitHub's CI
+- The number of elements requested by the Store Unit and the Element Requester now depends both on the requested eew and the past `eew` of the vector of the used register
+- When the VRF is written and `EMUL > 1`, the `eew` of all the interested registers is updated
+- Memory operations can change EMUL when EEW != VSEW
+- The LSU now correctly handles bursts with a saturated length of 256 beats
 
 ### Added
 
@@ -19,6 +23,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Vector single-width floating-point/integer type-convert instructions (`vfcvt.xu.f`, `vfcvt.x.f`, `vfcvt.rtz.xu.f`, `vfcvt.rtz.x.f`, `vfcvt.f.xu`, `vfcvt.f.x`)
 - Vector widening floating-point/integer type-convert instructions (`vfwcvt.xu.f`, `vfwcvt.x.f`, `vfwcvt.rtz.xu.f`, `vfwcvt.rtz.x.f`, `vfwcvt.f.xu`, `vfwcvt.f.x`, `vfwcvt.f.f`)
 - Vector narrowing floating-point/integer type-convert instructions (`vfncvt.xu.f`, `vfncvt.x.f`, `vfncvt.rtz.xu.f`, `vfncvt.rtz.x.f`, `vfncvt.f.xu`, `vfncvt.f.x`, `vfncvt.f.f`)
+- Vector whole-register move instruction `vmv<nr>`
+- Vector whole-register load/store `vl1r`, `vs1r`
+- Vector load/store mask `vle1`, `vse1`
+- Whole-register instructions are executed also if `vtype.vl == 0`
 
 ### Changed